# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NIPOPETO Pet Healthcare API",
    "version": "3.0.0",
    "description": "REST/JSON contract with bounded session/dashboard bootstrap and paginated collections. List GET response changed from array to {items,pagination,counts,related}. Default 20, maximum 100 rows."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/bootstrap": {
      "get": {
        "tags": [
          "Bootstrap"
        ],
        "summary": "Load session metadata or bounded dashboard previews",
        "operationId": "get_api_bootstrap",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bootstrap"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "session",
                "dashboard"
              ],
              "default": "session"
            },
            "description": "Session: profile, at most 20 owner pets and unread count. Dashboard: bounded previews and database totals; never full histories."
          }
        ]
      }
    },
    "/owners": {
      "post": {
        "tags": [
          "Owners"
        ],
        "summary": "Staff creates a counter customer profile without login credentials",
        "operationId": "createCounterOwner",
        "description": "Admin/staff only. Name and Vietnamese phone are required; email/address are optional. Normalizes +84 phones and detects existing contacts. Creates role owner with login disabled, returns no token or password. Does not change the staff session.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CounterOwnerInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Counter customer profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Admin/staff required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Existing phone/email, including a concurrent creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid contact information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "get": {
        "operationId": "listOwner",
        "summary": "List Owner",
        "tags": [
          "Pagination"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Page of Owner",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Owner"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/pets": {
      "get": {
        "tags": [
          "Pets"
        ],
        "summary": "List pets",
        "operationId": "get_api_pets",
        "responses": {
          "200": {
            "description": "Page of Pet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pet"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          },
          {
            "name": "species",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "Optional database filter. Use all or omit for no filter."
          },
          {
            "name": "healthStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "Optional database filter. Use all or omit for no filter."
          }
        ]
      },
      "post": {
        "tags": [
          "Pets"
        ],
        "summary": "Create pet",
        "description": "Owners create for their authenticated account. Staff supply ownerId for an existing user with role owner; missing owners or staff accounts are rejected with 422.",
        "operationId": "post_api_pets",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetInput"
              }
            }
          }
        }
      }
    },
...
    "/appointments": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "List appointments",
        "operationId": "get_api_appointments",
        "responses": {
          "200": {
            "description": "Page of Appointment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Appointment"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          },
          {
            "$ref": "#/components/parameters/PetIdFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "name": "statusGroup",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "pending (active), completed or cancelled (cancelled/no_show)."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "Appointments: medical or spa. Notifications: appointments, hotel, medical or promo."
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "Valid YYYY-MM-DD."
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "unbilled: eligible sources with no existing invoice."
          }
        ]
      },
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Create appointment",
        "operationId": "post_api_appointments",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentInput"
              }
            }
          }
        }
      }
    },
...
    "/medical-records": {
      "get": {
        "tags": [
          "Medical records"
        ],
        "summary": "List medical records",
        "operationId": "get_api_medical_records",
        "responses": {
          "200": {
            "description": "Page of MedicalRecord",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MedicalRecord"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          },
          {
            "$ref": "#/components/parameters/PetIdFilter"
          }
        ]
      },
      "post": {
        "tags": [
          "Medical records"
        ],
        "summary": "Create medical record",
        "operationId": "post_api_medical_records",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecordEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordInput"
              }
            }
          }
        }
      }
    },
...
    "/hotel-bookings": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "summary": "List bookings",
        "operationId": "get_api_hotel_bookings",
        "responses": {
          "200": {
            "description": "Page of HotelBooking",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HotelBooking"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          },
          {
            "$ref": "#/components/parameters/PetIdFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "name": "roomType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "Optional database filter. Use all or omit for no filter."
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "unbilled: eligible sources with no existing invoice."
          }
        ]
      },
      "post": {
        "tags": [
          "Hotel"
        ],
        "summary": "Create booking",
        "operationId": "post_api_hotel_bookings",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelBookingEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelBookingInput"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "Giữ nguyên key khi gửi lại cùng yêu cầu; cùng key và payload khác trả 409."
          }
        ],
        "description": "Tạo một yêu cầu lưu trú chưa thanh toán. Chống gửi lặp và trùng ngày cho cùng thú cưng."
      }
    },
...
  },
  "components": {
    "responses": {
...
      "ApiError": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Bootstrap": {
        "type": "object",
        "properties": {
          "currentOwnerId": {
            "type": "string"
          },
          "owners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Owner"
            }
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pet"
            }
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "medicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecord"
            }
          },
          "medicalImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalImage"
            }
          },
          "hotelBookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelBooking"
            }
          },
          "dailyCareNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareNote"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "summary": {
            "type": "object",
            "required": [
              "totals",
              "species",
              "appointmentStatus",
              "hotelStatus",
              "notificationCategory",
              "unread",
              "todayAppointments"
            ],
            "properties": {
              "totals": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "species": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "appointmentStatus": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "hotelStatus": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "notificationCategory": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "unread": {
                "type": "integer"
              },
              "todayAppointments": {
                "type": "integer"
              }
            }
          }
        },
        "required": [
          "currentOwnerId",
          "owners",
          "pets",
          "appointments",
          "medicalRecords",
          "medicalImages",
          "hotelBookings",
          "dailyCareNotes",
          "notifications",
          "summary"
        ]
      },
...
      "OwnerEnvelope": {
        "type": "object",
        "required": [
          "owner"
        ],
        "properties": {
          "owner": {
            "$ref": "#/components/schemas/Owner"
          }
        }
      },
...
    }
  }
}